extract vote button to own component, reuse in both articles and comments
